+++ espflash/src/cli/monitor/mod.rs
@@ -107,9 +107,13 @@ pub fn monitor(
     let firmware_elf = elfs.first().map(|v| &**v);
     let stdout = stdout();
     let mut stdout = if monitor_args.no_addresses {
+        if monitor_args.all_addresses {
+            log::warn!("Suppressing address resolution with `--all-addresses` given");
+        }
+
         ResolvingPrinter::new_no_addresses(firmware_elf, stdout.lock())
     } else {
-        ResolvingPrinter::new(elfs, stdout.lock())
+        ResolvingPrinter::new(elfs, stdout.lock(), monitor_args.all_addresses)
     };
 
     let mut parser: Box<dyn InputParser> = match monitor_args